Eliminate import cycles (tile types, unknown content, etc.)
