+++ src/models/tiles/placeholder/placeholder-content.ts
@@ -1,6 +1,6 @@
 import { types, Instance, SnapshotOut } from "mobx-state-tree";
 import { ITileModel } from "../tile-model";
-import { TileContentModel } from "../tile-types";
+import { TileContentModel } from "../tile-content";
 
 export const kPlaceholderTileType = "Placeholder";
