+++ src/models/tiles/tile-content.ts
@@ -1,28 +1,8 @@
 import { getEnv, getSnapshot, Instance, types } from "mobx-state-tree";
 import { SharedModelType } from "../shared/shared-model";
 import { ISharedModelManager } from "../shared/shared-model-manager";
-import { getTileContentModels, getTileContentInfo } from "./tile-content-info";
-import { ITileMetadataModel } from "./tile-metadata";
 import { tileModelHooks } from "./tile-model-hooks";
-
-/**
- * A dynamic union of tile content models. Its typescript type is `TileContentModel`.
- *
- * This uses MST's `late()`. It appears that `late()` runs the first time the
- * union is actually used by MST. For example to deserialize a snapshot or to
- * create an model instance. For this to work properly, these uses need to
- * happen after all necessary tiles are registered.
- *
- * By default a late type like this will have a type of `any`. All types in this
- * late union extend TileContentModel, so it is overridden to be
- * TileContentModel. This doesn't affect the MST runtime types.
- */
-export const TileContentUnion = types.late<typeof TileContentModel>(() => {
-  const contentModels = getTileContentModels();
-  return types.union({ dispatcher: tileContentFactory }, ...contentModels) as typeof TileContentModel;
-});
-
-export const kUnknownTileType = "Unknown";
+import { kUnknownTileType } from "./unknown-types";
 
 export interface ITileEnvironment {
   sharedModelManager?: ISharedModelManager;
@@ -88,56 +68,3 @@ export const TileContentModel = types.model("TileContentModel", {
   .actions(self => tileModelHooks({}));
 
 export interface ITileContentModel extends Instance<typeof TileContentModel> {}
-
-interface IPrivate {
-  metadata: Record<string, ITileMetadataModel>;
-}
-
-export const _private: IPrivate = {
-  metadata: {}
-};
-
-export function isRegisteredTileType(type: string) {
-  return !!getTileContentInfo(type);
-}
-
-export function tileContentFactory(snapshot: any) {
-  const tileType: string | undefined = snapshot?.type;
-  return getTileContentInfo(tileType)?.modelClass || UnknownContentModel;
-}
-
-export function findMetadata(type: string, id: string) {
-  const MetadataType = getTileContentInfo(type)?.metadataClass;
-  if (!MetadataType) return;
-
-  if (!_private.metadata[id]) {
-    _private.metadata[id] = MetadataType.create({ id });
-  }
-  return _private.metadata[id];
-}
-
-// The UnknownContentModel has to be defined in this tile-types module because it both
-// "extends" TileContentModel and UnknownContentModel is used by the tileContentFactory function
-// above. Because of this it is a kind of circular dependency.
-// If UnknownContentModel is moved to its own module this circular dependency causes an error.
-// If they are in the same module then this isn't a problem.
-//
-// There is a still an "unknown-content" module, so that module can
-// register the tile without adding a circular dependency on tile-content-info here.
-export const UnknownContentModel = TileContentModel
-  .named("UnknownContentModel")
-  .props({
-    type: types.optional(types.literal(kUnknownTileType), kUnknownTileType),
-    original: types.maybe(types.string)
-  })
-  .preProcessSnapshot(snapshot => {
-    const type = snapshot?.type;
-    return type && (type !== kUnknownTileType)
-            ? {
-              type: kUnknownTileType,
-              original: JSON.stringify(snapshot)
-            }
-            : snapshot;
-  });
-
-export interface IUnknownContentModel extends Instance<typeof UnknownContentModel> {}
