+++ src/models/tiles/unknown-types.ts
@@ -1,0 +1,1 @@
+export const kUnknownTileType = "Unknown";
